add login prototype
