+++ frontend/src/components/Login.js
@@ -21,16 +21,16 @@ const Login = () => {
     }
 
     const loginUser = async data => {
-        const fetchedUser = await fetch("http://locahost:4000/api/login", {
+        const fetchedUser = await fetch("http://localhost:4000/api/login", {
             method: 'post',
             headers: {
                 'Content-type': 'application/json',
                 'Accept': 'application/json'
             },
-            body: JSON.stringify(user),
+            body: JSON.stringify(data),
         })
 
-        return fetchedUser
+        return fetchedUser.json()
     }
 
     return (
