+++ frontend/src/components/Register.js
@@ -2,10 +2,30 @@
 
 const Register = () => {
 
+    const handleSubmit = async e => {
+
+    }
+
     return (
+        // <Form onSubmit={(e) => handleSubmit(e)}>
+        //     {error && <Form.Text className="text-danger mb-4">
+        //         Ungültige Kombination von Benutzername und Passwort!
+        //     </Form.Text>}
+        //     <Form.Group className="my-3" controlId="formBasicEmail">
+        //         <Form.Label>E-mail Addresse</Form.Label>
+        //         <Form.Control className="m-0" type="email" placeholder="E-Mail Adresse" onChange={(e) => setData({ ...data, email: e.target.value })}/>
+        //     </Form.Group>
 
-        <h1>REGISTER</h1>
+        //     <Form.Group className="mb-3" controlId="formBasicPassword">
+        //         <Form.Label>Passwort</Form.Label>
+        //         <Form.Control className="m-0" type="password" placeholder="Passwort" onChange={(e) => setData({ ...data, password: e.target.value })}/>
+        //     </Form.Group>
 
+        //     <Button variant="primary" type="submit">
+        //         Submit
+        //     </Button>
+        // </Form>
+        <>Register</>
     )
 
 }
